break out matrix utils
