+++ src/aspire/basis/dirac.py
@@ -3,8 +3,8 @@ import logging
 import numpy as np
 
 from aspire.basis import Basis
+from aspire.utils import roll_dim, unroll_dim
 from aspire.utils.matlab_compat import m_flatten, m_reshape
-from aspire.utils.matrix import roll_dim, unroll_dim
 
 logger = logging.getLogger(__name__)
