+++ src/aspire/basis/fb_2d.py
@@ -5,9 +5,8 @@ from scipy.special import jv
 
 from aspire.basis import Basis
 from aspire.basis.basis_utils import unique_coords_nd
-from aspire.utils import ensure
+from aspire.utils import ensure, roll_dim, unroll_dim
 from aspire.utils.matlab_compat import m_flatten, m_reshape
-from aspire.utils.matrix import roll_dim, unroll_dim
 
 logger = logging.getLogger(__name__)
