+++ src/aspire/covariance/__init__.py
@@ -1,2 +1,2 @@
 from .covar import CovarianceEstimator
-from .covar2d import BatchedRotCov2D, RotCov2D
+from .covar2d import BatchedRotCov2D, RotCov2D, shrink_covar
