+++ src/aspire/covariance/covar.py
@@ -12,15 +12,15 @@ from aspire import config
 from aspire.image import Image
 from aspire.nufft import anufft
 from aspire.reconstruction import Estimator, FourierKernel, MeanEstimator
-from aspire.utils import ensure
-from aspire.utils.fft import mdim_ifftshift
-from aspire.utils.matrix import (
+from aspire.utils import (
+    ensure,
     make_symmat,
     symmat_to_vec_iso,
     vec_to_symmat_iso,
     vecmat_to_volmat,
     volmat_to_vecmat,
 )
+from aspire.utils.fft import mdim_ifftshift
 from aspire.volume import Volume, rotated_grids
 
 logger = logging.getLogger(__name__)
