+++ src/aspire/image/image.py
@@ -8,11 +8,10 @@ from scipy.linalg import lstsq
 
 import aspire.volume
 from aspire.nufft import anufft
-from aspire.utils import ensure
+from aspire.utils import anorm, ensure
 from aspire.utils.coor_trans import grid_2d
 from aspire.utils.fft import centered_fft2, centered_ifft2
 from aspire.utils.matlab_compat import m_reshape
-from aspire.utils.matrix import anorm
 
 logger = logging.getLogger(__name__)
