+++ src/aspire/reconstruction/kernel.py
@@ -3,16 +3,16 @@ import logging
 import numpy as np
 from scipy.fftpack import fft, fftn, fftshift, ifft, ifftn
 
-from aspire.utils import ensure
-from aspire.utils.fft import mdim_fftshift, mdim_ifftshift
-from aspire.utils.matlab_compat import m_reshape
-from aspire.utils.matrix import (
+from aspire.utils import (
+    ensure,
     roll_dim,
     unroll_dim,
     vec_to_vol,
     vecmat_to_volmat,
     vol_to_vec,
 )
+from aspire.utils.fft import mdim_fftshift, mdim_ifftshift
+from aspire.utils.matlab_compat import m_reshape
 
 logger = logging.getLogger(__name__)
