+++ src/aspire/source/simulation.py
@@ -7,9 +7,8 @@ from aspire.image import Image
 from aspire.image.xform import NoiseAdder
 from aspire.operators import ZeroFilter
 from aspire.source import ImageSource
-from aspire.utils import ensure
+from aspire.utils import acorr, ainner, anorm, ensure, make_symmat, vecmat_to_volmat
 from aspire.utils.coor_trans import grid_3d, uniform_random_angles
-from aspire.utils.matrix import acorr, ainner, anorm, make_symmat, vecmat_to_volmat
 from aspire.utils.random import Random, rand, randi, randn
 from aspire.volume import Volume
