+++ src/aspire/utils/__init__.py
@@ -1,2 +1,24 @@
+from .matrix import (
+    acorr,
+    ainner,
+    anorm,
+    eigs,
+    im_to_vec,
+    make_symmat,
+    mat_to_vec,
+    mdim_mat_fun_conj,
+    roll_dim,
+    symmat_to_vec,
+    symmat_to_vec_iso,
+    unroll_dim,
+    vec_to_im,
+    vec_to_mat,
+    vec_to_symmat,
+    vec_to_symmat_iso,
+    vec_to_vol,
+    vecmat_to_volmat,
+    vol_to_vec,
+    volmat_to_vecmat,
+)
 from .misc import ensure, get_full_version, powerset, sha256sum
 from .types import complex_type, real_type, utest_tolerance
