+++ tutorials/examples/cov3d_simulation.py
@@ -14,7 +14,7 @@ from aspire.noise import WhiteNoiseEstimator
 from aspire.operators import RadialCTFFilter
 from aspire.reconstruction import MeanEstimator
 from aspire.source.simulation import Simulation
-from aspire.utils.matrix import eigs
+from aspire.utils import eigs
 from aspire.utils.random import Random
 from aspire.volume import Volume
